ASIdiSPIM: handle changes in micro-mirror (scanner) firmware v3.14
git-svn-id: https://valelab.ucsf.edu/svn/micromanager2/trunk@15968 d0ab736e-dc22-4aeb-8dc9-08def0aa14fd

+++ plugins/ASIdiSPIM/src/org/micromanager/asidispim/Utils/SliceTiming.java
@@ -32,7 +32,7 @@ package org.micromanager.asidispim.Utils;
 public class SliceTiming {
    public float scanDelay;
    public int scanNum;
-   public int scanPeriod;
+   public float scanPeriod;
    public float laserDelay;
    public float laserDuration;
    public float cameraDelay;
